reducers folder renamed to hooks
