+++ app/javascript/hooks/SearchReducer.js
